[REM] mrp_workorder: instruction in barcode menu
Before the OWL conversion of the `stock_barcode` main menu, the override
of its template to add the Manufacturing instruction doesn't cause any
troubles even if `stock_barcode` wasn't installed (`mrp_workorder`
depends of `stock` and `barcodes` but not `stock_barcode`).
But after the OWL conversion, the override raises an error if
`stock_barcode` isn't installed.
Instead of creating a new module only to add a sentence in a XML
template, we remove it, since the PDF with the instructions still can be
downloaded in the settings.

closes odoo/enterprise#31009

Signed-off-by: William Henrotin (whe) <whe@odoo.com>