[IMP] planning: improve user experience
This commit removes the tags from the planning slots and adds a role field
to the material resource. The objective is to fulfill some specific use
case. e.g. : you need 2 cranes for your construction. You can create open
shifts for the two following roles: crane 1 and crane 2.

We remove the planning_calendar module which was done as part of task 2972953.
This change was made as part of the general effort to simplify the planning app
and improving its user friendliness.

This commit adds a menu option under the "Configuration" option that redirects
the user to the employees kanban/list view.

This change will allow the user to view the employees straight from the planning
app, further increasing efficiency and ease of use.

This commit changes the defaults hours that are printed in the shift form view
from 00:00-23:59 to 09:00-17:00 which are the standard working hours.
This only applies to non-working days, where the employee who is to be assigned
a shift does not have standard working hours.

This change should make it easier for managers to assign shifts on non-working days.

It also rewords a few onboarding tours to be more accurate

task-2972953

closes odoo/enterprise#31172

Related: odoo/upgrade#3877
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Co-authored-by: Panagiotis Kyriakou <paky@odoo.com>