[FIX] account_reports: foldable section
This fix allows you to fold a line that has sublines.
Indeed, before this fix, putting a foldable field on this kind of line did not change anything when the line had no result and therefore did not allow to fold it.

Thanks to this fix, all sections (so a line with other lines as children) will be foldable.

closes odoo/enterprise#31022

Signed-off-by: maximilien(malb) <malb@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>