[IMP] planning: show a preview of the gantt as role sample data
Previous to this commit the helper displayed for the role was the regular one.
In order to improve the onboarding, a preview of the gantt grouped by role and
resource will be displayed.

This feature as been limited to non portable devices.

As it was part of the requirements not to use and image and rather use html, the
code has been minified in order not to bloat the file.

task-2971483

closes odoo/enterprise#31031

Signed-off-by: Xavier <xbo@odoo.com>