[FIX] web_enterprise: restore old color of list view title
Before in BS4, the table used the `$table-head-color` SCSS variable to
color the header of table. This variable doesn't exist anymore on BS5.
In the commit of the BS5's merge, we have replaced the old value by
a value calculated with the color contrast in the case of the list view
it wasn't the right fix, so in community commit we restore the original
value like it was before the merge of BS5.

In this commit, we remove the unused line that was left in
the migration of BS5.

Note:
In BS4 branch the result is `#212529`:
```scss
$o-gray-900: #212529 !default;
$o-main-headings-color: $o-gray-900 !default;
$table-head-color: $o-main-headings-color !default;
.o_list_view .o_list_table thead {
  color: $table-head-color;
}
```
In BS5 branch the result is `#00000`:
```scss
.o_list_view .o_list_table thead {
  color: color-contrast(opaque($body-bg, $light));
}
```

closes odoo/enterprise#31026

Related: odoo/odoo#99488
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>