[IMP] documents: improve on search panel edit
This commit is another iteration on the edit from search panel feature
in the documents app.
Instead of having to double click on the folder/tag the plus has been
replaced with a cog which will open a popover with multiple options like
edit create and delete.

TaskId-2971448

Part-of: odoo/enterprise#31038