[IMP] planning,hr_work_entry_contract_planning: manage flexible hours
Before this commit, in Planning, the shift's allocated hours are
automatically calculated based on the working calendar of the resource
except if the shift lasts <24h.
It was implemented this way to accommodate the use-case of people with
flexible hours e.g. nurses, firefighters, waiters etc. And so if the
working calendar were takken into account, night shifts would result
in 0 allocated hours (which it's wrong for the people with a flexible
hours). The side effect of the current implementation is that allocated
hours are not "correctly" calculated for people working with fixed
hours (that is people who follows a working schedule) because if a
shift is created from 8 AM to 5 PM, the allocated would be 9 allocated
hours instead of the expected 8 hours according to the working schedule
set on the resource linked to the shift (In other words, the typical
"lunch break" isn't deducted from the calculation).

This commit adds a flexible hours boolean field on the
`resource.resource` and `hr.employee` models. When the field is True,
the allocated hours computed will not take into account the working
schedule of the resource even if the period of the shift is longer than
one day, however, the max allocated hours are the hours_per_day set on
the working schedule set on the company set on the shift.
When the flexible hours is False than the allocated hours
computed will take into account the working schedule of the resource
linked to the shift or the one set on company set on the shift if no
resource and it is longer than one day.

task-2971572

closes odoo/enterprise#31110

Related: odoo/odoo#99693
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>