[ADD] appointment: onboarding
**Purpose**
Add an 'Onboarding' to Appointment

**Notes**
See the COM PR for the concurrent implementation of the new onboarding module.

The onboarding adds three steps:
1/ Create an appointment type.
If one was already created, this one will be opened for confirmation.

2/ Create an appointment invite.
If no appointment existed, create a simple one on-the-fly.
In more details,
The preview link step is specialized vs. appointment.invite's forw view to:
- Use appointment name for link, allowing to reopen it and reuse without error
- validate preview link onboarding step state only after one of the buttons is
pushed

3/ Configure the external calendar provider for synchronization.
We added this logic to the calendar module and use it here as well.

This onboarding is added to the kanban/list and list views of Online
Appointments, as well as the Calendar view as it is the gateway into the
appointment module-specific pages.

Task-2852375
Part of Task-2900763
See odoo/odoo#97105

Part-of: odoo/enterprise#29139
Co-authored-by: Florian Charlier <flch@odoo.com>
Co-authored-by: Jeremy Hennecart <jeh@odoo.com>