[REF] web_enterprise: remove props of ActionSwiper
This commit removes the isActive props, used as a workaround
because Owl didn't supported multiple occurence of the same
slot in different locations. Since odoo/odoo#99491, the issue
has been fixed and the component no longer require this
props to let the settings only activate the swiper on smaller
screens.

closes odoo/enterprise#31078

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>