[FIX] mrp_mps: include interwh delay
To reproduce the issue:
(Let WH01 be the main warehouse)
1. Create a second warehouse WH02
2. Edit WH01:
    - Resupply From: WH02
3. Edit the resupply route of WH01:
    - Edit the rule from InterWH Transit to WH01/input
        - Lead time: 1
4. Create a storable product P:
    - Type: Storable
    - Routes: "WH01: Supply... from WH02"
5. Add P on MPS view and enable the row "Actual Replenishment"
6. Set a positive replenish quantity for P (not in the first interval)
7. Replenish P

Error: The actual replenishment is updated on the previous interval. It
should be on the interval defined at step 6.

When computing the date of an incoming stock move, we get and use the
delay of its dest moves but we don't consider the delay of the move
itself.

Moreover, `_get_dest_moves_delay` handles the case where a SM does not
have any dest move, so `_get_moves_and_date` can be simplified

OPW-2906741

closes odoo/enterprise#31287

X-original-commit: b67903092d423913530c024eb2dbc40ffa30e98b
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>