[FIX] stock_barcode: opening a transfer with several companies selected
Steps to reproduce the bug:
- Have companies A and B
- Select company A as current company and B as allowed company
- Go to barcode app > Operations > Select receipts of the company B
- Try to create a new picking

Problem:
An access error is triggered because we access company B data while only
company A is in the allowed companies.

When we get the record `stock.picking`, we make sure that the value of
`env.company` is correct with the `with_company`, but this function will
also redefine the authorized companies with the context key
`allowed_company_ids`, Except that this key is not present in the
context of the record. So`with_company` will define "company A" as the
current company (correct) but as the only authorized company
(incorrect):
https://github.com/odoo/odoo/blob/4c611efa959f8b856ef4ba1a16df664acd0cfecf/odoo/models.py#L5283-L5293

Instead, we can define the allowed companies with the `with_context`
function. The ORM will be able to deduce the current company:
https://github.com/odoo/odoo/blob/e902713648bca329e6821859462b2128aad62c09/odoo/api.py#L639

opw-2926638

closes odoo/enterprise#31323

X-original-commit: 04934cde3f38f689708dc7c789c98ccb3ac25b65
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>