[FIX] l10n_mx_reports: allow opening DIOT report
DIOT report was failing because we were building the result dict of the engine based upon the column_ids' expression_label. This was wrong, because expression labelled 'counter' wasn't included, since it does not correspond to any column. This expression is necessary for exports, though, so instead of calling the columns, we now directly call the labels of the expressions defined on the report line.

closes odoo/enterprise#31210

Signed-off-by: Laurent Smet <las@odoo.com>