[IMP] hr_payroll: allow Time Off Officer to do second approval
As part of https://github.com/odoo/odoo/pull/99496, buttons to approve were made available in the Time Off tree view.
However, the validate button requires to recompute payslips. The Time Off officer has no right to do it so it was giving an error.

This commit fixes the access right to allow the Time Off Officer to do a second approval of the leaves.

task-2939201

closes odoo/enterprise#31085

Related: odoo/odoo#99496
Signed-off-by: Kevin Baptiste <kba@odoo.com>