[FIX] hr_payroll_holidays: fix traceback on create payslip
After https://github.com/odoo/enterprise/pull/30468, it was decided to show the
time off to defer warning in the form view only if the related employee of the
payslip had some time off to defer. However, when creating a payslip, we don't
have an employee yet and a traceback is shown. This commit reverts this
behavior. So if there is a time-off to defer, the warning will be shown
everytime.

task-2976064

closes odoo/enterprise#31109

Signed-off-by: Kevin Baptiste <kba@odoo.com>