[IMP] account: cash discount improvements
Purpose :
Improve multiple use cases while handling the Accounting part of Cash Discounts in Odoo.

When creating a Payment Term, the user can now specifically design an early payment discount through an additional part in the wizard.
This will, amongst other things, allow the user to choose in which way the tax should be computed on the discounted invoice (tax included, excluded, or tax base impacted).
(WIP : belgian taxation - Impact Tax Base - isn't yet functional. At terms, the taxation of the whole invoice should be adapted when an Cash Discount is st within a payment term. This option should only be available on the belgian localization. )

The discounted amount and the conditions to fulfill are printed below the invoice PDF.

Registering a payment while filling the condition for an early payment discount will prefill the fields accordingly.

Reconciliation --> In the reconciliation widget, the logic of Early Payments Discounts are also applied.

Related : odoo/odoo#94050

closes odoo/enterprise#31225

Task: 2968644
Related: odoo/odoo#99572
Signed-off-by: Laurent Smet <las@odoo.com>