[FIX] sign: recreate controller to allow old share link to be signed
Before this commit, the shared templates that were shared in 15.2 and older
could not be signed anymore. The controller was deleted. If a parter received
the mail before the upgrade and wanted to use it, it would return a 401 error.

This commit put back the controller. The data are updated in a upgrade script
where the corresponding sign request are created and the old share_link value
is stored in the access_token field.

Upgrade PR:

closes odoo/enterprise#31177

X-original-commit: b91d7399a28cb0b1aa2bddf9eb950494307f6fa0
Signed-off-by: Jérémy Kersten <jke@odoo.com>
Signed-off-by: Arnaud Joset <arj@odoo.com>