[FIX] knowledge: allow to create article items from hierarchy tree
Before this commit, when trying to create a child article on an article that
has article item children, there was an error raised saying that we cannot
mix both types of articles (normal and item) under the same parent.

After this commit, the "quick create" detects that the parents has item
children and set the 'is_article_item' flag on the article to create.
This allow to create article items directly from the side panel.

Task-2858661

X-original-commit: 79f81c7f864f994d3ed5b9507b65a2a3072c4ec7
Part-of: odoo/enterprise#32070