[FIX] knowledge: fix display of portal side panel
Before this commit, if a private root article was published, the side panel of
the portal view was displayed, even if there was no article to display. This
because the show_side_panel flag was only based on if there was "avaible"
root articles, no matter their category. The private articles are not used at
all in the side panel tree.

After this commit, the flag is based on root article that are not private.
So that the side panel is only displayed if there are some article to display
in it.

Task-2858661

Part-of: odoo/enterprise#31852