[FIX] industry_fsm{_report}: make the tour independant of demo data
Prior to this commit, the tour was using a field that is generated from the demo data,
which is not right. Indeed, the `x_description` field is created from the demo data, so
it is not found during the tour when no demo data is installed and the tour fails.

task-2973342

Part-of: odoo/enterprise#31087
Co-authored-by: Prakash Prajapati <ppr@odoo.com>