[FIX] documents: fix traceback when switching from workspace
The documents app tries to keep the selected records when switching from
views/workspace however it was badly handled in the case of grouped
views (as it was mainly designed with the kanban in mind) leading to a
stack trace.
This commit ensures that the traceback does not happen.

TaskId-2971448

closes odoo/enterprise#31038

Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>