[FIX] knowledge: make favourite list clickable on add/remove favourite
Currently, in the knowledge module, when a portal user adds/removes an article
to their favorite list, the whole list is unclickable.

To fix the issue, we have added `portal_readonly_mode` param in the template
render according to the user rights and it will show a clickable link if the
user has correct rights.

task-2925871

closes odoo/enterprise#31158

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>