[FIX] sale_subscription: SO without subscription cancellation
It seems we can have SOs in a situation where they have their
`subscription_management` set to either `upsell` or `renew`, but without
having a parent `subscription_id`. This is causing an issue in the SO
cancellation flow because it would try to post the cancellation message
in the parent subscription's chatter. This is crashing because there's
no subscription to write on.

via p/feedback

closes odoo/enterprise#31182

X-original-commit: 35528ffec49d79ebee37194982aa2a722f027551
Signed-off-by: Arnaud Joset <arj@odoo.com>