[FIX] pos_settle_due: Open the POS without accounting rights
When you have some POS access rights and no accounting rights, you are
not able to open the POS session because it tries to get the field total
due on the partners, but does not have the required access rights.

To avoid this issue, we are not loading the total due if the user does
not have the access rights.

closes odoo/enterprise#31197

X-original-commit: b2df0fb20667da9e73bacf302dd054dee2154995
Signed-off-by: Masereel Pierre <pim@odoo.com>