[FIX] sale_subscription: can not archive partner or employee
Steps to reproduce:
- Install hr and subscription
- Set no subscription rights and admin hr rights on a user
- Try to archive an employee

Current behavior:
Error for access rights to sale.subscription

Expected behavior:
No error

Explanation:
when we write the active to false of the res.partner
(corresponding to the employee.address_home_id) there is
an issue as you need the rights to search and modify the
sale subscritpions. You should not need the rights for
this access to sale.subscription hence we add a sudo to
solve the issue.

opw-2974590

closes odoo/enterprise#31519

X-original-commit: bd78fabbdc724adf4c7dd6d8934289db98a5b0c8
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Fockedey Martin (mafo) <mafo@odoo.com>