[FIX] documents: handle False default values when uploading documents
When a default value in the context was set to False in the documents kanban view, upon uploading a document, the value sent to the 'portal_my_task_document_upload' route for that field was the string 'false'.
With this commit, the default values set to False are now ignored and not sent at all to the back end.

Community: odoo/odoo#99151
Upgrade: odoo/upgrade#3749

Task-2944247

Part-of: odoo/enterprise#30838