[IMP] knowledge: reduce hovering zone of icon/cover buttons
Purpose:
========
Make the area that makes the "add cover" and "add button" appear smaller.

Previously, these buttons where shown when hovering the body, which made
them appear too often when the user did not plan to use them.
Now, the buttons are shown when the cursor is near them (between the body
and the header).

Task-2852916

Part-of: odoo/enterprise#31744