[FIX] stock_barcode: constraint
In the following commit: dd7f4d501a4d23b04a1899be8fd51cc583673c9d
- We had new fields for `stock.picking.type`;
- We had a method who check these fields's value in the `create` and in
  the `write`;
- We force `restrict_scan_source_location` to `mandatory` in the data
  for the outgoing picking type.

Those cause an issue: when setting the warehouse's outgoing shipments
setting in 3 step (pick-pack-ship), the check method raises the error.
It was because the `_create_or_update_sequences_and_picking_types`
warehouse method call `update` for every fields one by one, meaning when
we want to update `restrict_scan_source_location` and
`restrict_scan_dest_location`, it write first the
`restrict_scan_source_location` which will raise the error because it
doesn't update yet `restrict_scan_source_location` (and so, we have the
forbidden couple "mandatory"/"optional").
Maybe we should call `write` instead of `update` in the warehouse's
method, but that said, this commit makes the check method a python
constraint (instead of calling it in `write`/`create`).

Also, to do that, we have to force 'restrict_scan_dest_location' on 'no'
for the outgoing picking type otherwise the constraint is violated by
the data and can't be added.

closes odoo/enterprise#31642

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>