[IMP] industry_fsm,project_forecast: no open project in list view
Before this commit, the project field in task and
planning.slot{.template} models is clickable in list view of those
models.

This commit prevents to open the form view in the list view since the
form view of project is now more a view to configuration the project and
so once the configuration is done, the user will not always need to go
to that view.

closes odoo/enterprise#31212

Related: odoo/odoo#99949
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>