[ADD] website_sale_subscription : eCommerce for subscription
This module allows you to sell subscription products in
your eCommerce with appropriate views and selling choices.

This module has for goal to fix the behaviour of recurring
product in eCommerce.
This module mainly add the recurring information to price tags,
create a subscription SO, and disallow adding multiple items
with different 'first_suitable_pricing's recurrence.

task-id : 2826815

closes odoo/enterprise#31232

Related: odoo/odoo#100210
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>