[ADD] website_sale_subscription : eCommerce for subscription
This module allows you to sell subscription products in
your eCommerce with appropriate views and selling choices.

This module has for goal to fix the behaviour of recurring
product in eCommerce.
This module mainly add the recurring information to price tags,
create a subscription SO, and disallow adding multiple items
with different 'first_suitable_pricing's recurrence.

task-id : 2826815

closes odoo/enterprise#32716

X-original-commit: 736a026e9453555e20336c7b0d3009e7cc181ccc
Related: odoo/odoo#103078
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>