[FIX] account_3way_match: apply better perf trick for can_be_paid
Using UPDATE to speed up module installation is not the best way to do it,
because that touches every single line.

Instead add the column with a DEFAULT value in the database schema directly.
That way PostgreSQL will not need to touch the rows at all, the operation will
be instantaneous, and the actual values will be lazily initialized when the rows
are modified later.

Previous implementation: c426a4532d507b729f677f220b4571a096d188d5

This commits applies the same trick for field `account.move:release_to_pay`

opw-2894901

closes odoo/enterprise#31494

X-original-commit: a22dac22d635ad4d25ff751dd376dab3cbaf3344
Signed-off-by: William André (wan) <wan@odoo.com>