[IMP] knowledge: use non-legacy emoji picker in knowledge
This commit replaces the legacy emoji picker by the new
emoji picker in knowledge.

The emoji picker is triggered by the same user interactions
as the legacy emoji picker, i.e. click on an icon in a
knowledge article.

Knowledge emoji picker has the default behaviour of base
emoji picker, and then slightly extends with a action to
remove the icon of emoji. This remove action is a button
in the header part of emoji picker.

Task-2980528

closes odoo/enterprise#31231

Related: odoo/odoo#100010
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>