[FIX] delivery_easypost: test env easypost-ups for packages
Steps to reproduce:
1) Create Shipping Methods:
- put a test and a production key
- Carrier Type: UPS (you have to load them beforehand with the BUTTON)
- Default package: "Pak"

2) Create new product:
- Weight = 1
- Storable Product
- On hand product : 25 units

3) Create new product:
- Weight = 2
- Storable Product
- On hand product : 25 units

4) Create a sale order:
- Add the two created products
- Add a shipping Easypost - click on get rate before adding it
- confirm

5) Click on the smart button delivery
6) Set the quantity done for one of the product and click on "PUT IN PACK" - select "Pak"
7) Set the quantity done for the other product and click on "PUT IN PACK" - select "Pak"
8) Validate

Issue:
- Traceback: 'NoneType' object is not subscriptable

Cause:
In a test environment, with easypost-ups, whenever you have only one `tracking_id` for multiple shipments (by default in test env),
easypost-ups will respond with only one `tracker` for the first shipment (package) and set the `tracker` for the other shipments (packages) to Null.

Transcript with easypost tech support:
```
I did some testing on my end and found that there will be only one tracker object returned when in the Test environment with UPS. I also found that in production, I received a tracker object as well as tracking URLs for both shipments in the order.

You are correct with your original outreach, that in test, you would only receive one tracker object since the tracking codes are the same.
```
Solution:
Though not beautiful, when in test mode and with UPS, select only the `tracker` for the first shipment.

opw-2963499

closes odoo/enterprise#31535

X-original-commit: 283c5851ad6a8e9f3d562f7682bfa8d0d23ce9f6
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Yolann Sabaux (yosa) <yosa@odoo.com>