[FIX] quality_control: cancel PO with basic access
When canceling a PO, if the user only has an access to Purchase, an
access error will be triggered: "Sorry, you are not allowed to delete
documents of type 'Quality Check'"

There are actually two changes that are needed here:

Fix in SM:
Cancelling a PO leads to the SM cancellation. At that step, we check if
we need to unlink some QC (`/quality_control:StockMove._action_cancel`).
To avoid any access issue, we consult all related QC in sudo mode and
each time a QC must be deleted, we append it to an existing recordset
`to_unlink`. But here is the issue: `to_unlink` has not been created in
sudo mode. Therefore, even if the added records have the `su` flag, the
recordset itself does not have it as it keeps its initial environment.

More generically, if `a` and `b` are two existing records:
```py
(a | b.sudo()).env.su == False
(a.sudo() | b).env.su == True
```

Fix in SML:
Cancelling a PO leads to the SM unreservation and possibly the SML
deletion. Such an operation also leads to the related QC deletion. This
may trigger an access error if the user does not have any access to
Quality.

closes odoo/enterprise#31300

X-original-commit: c825128462651467155f4f26d8fb00d043c72b62
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>