[FIX] web: list selection banner takes the whole width on mobile
On mobile, in list view's selection mode, both the left & right bottom
parts of the ControlPanel are displayed but only the left one
(containing the selection banner and actions menu) should be displayed
(and take the full width).

This commit adapts tests accordingly to the community changes.

closes odoo/enterprise#31269

Related: odoo/odoo#99483
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>