[IMP] knowledge: reduce space between emoji/buttons and body
Purpose:
========
Reduce the empty space between the "Add icon" button and the article
body, and between the article emoji and the body.
Removed the highlight of the article emoji on hover (which seemed to be
intended for emojis in the sidebar), because with the space reduction,
it overlapped the title.

Task-2852916

Part-of: odoo/enterprise#31744