[LINT] mail, various: use explicit subtype in message_post_{with_...}
Purpose of this commit is to be explicit in subtype chosen when invoking the
message composer / calling message_post. As default value may not always be
clear, better be explicit in case the composer default value changes.

Some code cleaning is also performed as the same time, as done in community.

See community PR and related commit for more details.

Task-2710804 (Mail: Clean MailThread Posting API)

Part-of: odoo/enterprise#31017