[IMP] stock_barcode: display +1 button for sn
Before this commit, the button to add the reserved quantity (or +1 if
it's not a planned move line) was not shown for the line with a product
tracked by serial numbers.
It will now be displayed if a serial is reserved, if the quantity done
is equals to 0 and if the scan of the tracking number isn't mandatory.

task-2960938

X-original-commit: 48b7041019ec6e7c265ee41697d036efaf1b6bd8
Part-of: odoo/enterprise#32481