[IMP] account_intrastat: demo data
This commit adds intrastat commodity codes for the demo data
'consu' (i.e. not service) products present in the product
module (intrastat depends on accounting, which depends on product, thus
we can add to these products). Only one of these products requires data
for the intrastat_supplementary_unit_amount field.

This relates to bugprovements for v16:
https://pad.odoo.com/p/accounta_matata_v16

closes odoo/enterprise#32548

X-original-commit: bf0dac1a19a0583bef68f6e95ed0a995b39f5019
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>