[IMP] account_intrastat: demo data
This commit adds intrastat commodity codes for the demo data
'consu' (i.e. not service) products present in the product
module (intrastat depends on accounting, which depends on product, thus
we can add to these products). Only one of these products requires data
for the intrastat_supplementary_unit_amount field.

This relates to bugprovements for v16:
https://pad.odoo.com/p/accounta_matata_v16

closes odoo/enterprise#32379

Signed-off-by: Josse Colpaert <jco@odoo.com>