[IMP] hr_appraisal: prevent activity creation on appraisal reopening
Currently, when a manager wants to reopen an appraisal to correct a
small mistake, the corresponding activities are recreated. This commit
prevents the creation of these activities when an appraisal goes from
the 'done' stage back to the 'confirmed' state. To ensure that people
can't just silently make changes to appraisals this way after the fact,
followers will be notified when appraisals are set to 'done'.

task-2983617

closes odoo/enterprise#31305

Signed-off-by: Kevin Baptiste <kba@odoo.com>