[IMP] hr_appraisal: prevent activity creation on appraisal reopening
Currently, when a manager wants to reopen an appraisal to correct a
small mistake, the corresponding activities are recreated. This commit
prevents the creation of these activities when an appraisal goes from
the 'done' stage back to the 'confirmed' state. To ensure that people
can't just silently make changes to appraisals this way after the fact,
followers will be notified when appraisals are set to 'done'.

task-2983617

closes odoo/enterprise#31961

X-original-commit: c5f31aa0a603710cecdf6bba5a19d5e471a623f9
Signed-off-by: Kevin Baptiste <kba@odoo.com>