[FIX] industry_fsm: prevent undeterministic error in tour
Prior to this commit, some of the `trigger` on a many2one field
where made on `ul.ui-autocomplete > li:first-child > a` which could
lead to clicking on the loading if the rpc was not yet finished.

This commit ensure that the `loading` is not taken by the tour.

closes odoo/enterprise#31309

Related: odoo/odoo#100125
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>