[IMP] iot: iot box restart Odoo + reboot buttons
Before, the user had to manually unplug and replug the IoT Box to reboot it.
Also, it was required to connect to the IoT box to restart Odoo service.

This improvement adds the 'Restart Odoo' and 'Reboot IoT box' buttons on the box's form view to IoT app.
They allow the user to reboot the IoT box and restart Odoo service on it remotely from the IoT app in Odoo.

task-2476576

closes odoo/enterprise#31723

Related: odoo/odoo#100999
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>