[FIX] sign: broken auto completion
Following d3ea349, the values provided in the evaluation context for
auto-fields is in a key named `auto_value`, to avoid confusion with
the name of the field itself.

However, a bugfix in 15.0 was forward-ported with the outdated
naming in 0f42d1d, which broke the auto-field auto-filling.

This commit sets the key to its appropriate value to restore the
functionnality.

closes odoo/enterprise#31461

X-original-commit: 6d17e0b4fa8cbd2d6550138f0784a20e02cf6e30
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Bouvy Damien (dbo) <dbo@odoo.com>