[FIX] web_studio,*: find correct xpath of field
*: industry_fsm_report, worksheet

First Issue
===========
To Reproduce
------------
- Enable Analytic tags on Accounting settings
- open a sale order
- open studio
- select edit list view on sale order lines
- show hiden elements and select Analytic Tags
- check *Disable Creation*

A Validation Error is raised

Problem
-------

The argument `model` in the overrides of the method `edit_view` in `industry_fsm_report` and `worksheet` is missing,
When the `edit_view` is called in `web_studio` the argument will be `None` then will be replaced by `view.model`
https://github.com/odoo/enterprise/blob/203eafdf2cdaab034c83c628d5e81d0265b66860/web_studio/controllers/main.py#L566
which is `sale.order` in this case.
then at this line :
https://github.com/odoo/enterprise/blob/203eafdf2cdaab034c83c628d5e81d0265b66860/web_studio/controllers/main.py#L1036
we will look for `analytic_tag_ids` in the model `sale.order` which can't be there and that what leads to the error.

Solution
--------

To solve the issue we add the argument `model` to the overrides of `edit_view`

Second Issue
============
To Reproduce
------------
Having the first issue fixed, when we check *Disable creation* on Analytic Tags, it's automatically unchecked and it
doesn't modify anything.

Problem
-------
the form view of `sale.order` has a `form` and a `tree`, the filed `analytic_tag_ids` is present in both views, so when
we apply the changes on `analytic_tag_ids` in the `tree` view, because of not precising the correct xpath we take the first
field `analytic_tag_ids` which is the one of the `form`, and that explains this behavior.

Solution
--------
the problem comes from this weird/extraordinary condition :slightly_smiling_face: :
https://github.com/odoo/enterprise/blob/203eafdf2cdaab034c83c628d5e81d0265b66860/web_studio/controllers/main.py#L821
where I think the writer of it wanted to simply do :
```py
if expr not in xpath:
	expr = xpath + expr
```

opw-2978291

closes odoo/enterprise#31360

closes odoo/enterprise#34699

X-original-commit: 6c15043dd0fa4715a8e5ea49b93cc8f0ca8ec798
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>