[IMP] knowledge: display favorite children in sidebar
This commit allows to display the favorites children in the side bar. That way
users can favorite an article header and have access to its children directly
from the favorite sidebar.

Drag and drop of those children is not allowed. Drag and drop under or between
those children is not allowed.

Task-2985026

Part-of: odoo/enterprise#31890