[FIX] l10n_be_hr_payroll: Fix double holiday pay on attests
Purpose
=======

If June has passed and the payment of the double has already taken
place at the company, there is in principle no recalculation of the
double pay at all on the N-1 holiday form (on the other hand, we
must calculate single/double /complementary on sheet N)

Specification
=============

The double holiday pay shouldn't be computed on the ***N-1***
holiday attests if the double holiday pay has already been paid
the current year.

The N is computed as before.

closes odoo/enterprise#31404

X-original-commit: 7e4549e0afaa914cb87decca5efbd8e097f45ab0
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>