[FIX] knowledge: disable drag and drop on Open article
Before this commit, this was possible to start dragging an article of the side
panel even after user clicked to open an article. This led to an error because
after reloading, dragging is still active. When dropping after reload of the
page, the sortable component of the previous page is in fact destroyed.

After this commit, the sortables are disabled when clicking on an article to
open it. User cannot start dragging an article while the next article to display
is loading.

Task-2972864

X-original-commit: 3509b39c9f90b61ca471f7e06799493949b506fb
Part-of: odoo/enterprise#31700