[FIX] hr_contract_salary: restore old styling for Select2
Due to the migration of Bootstrap 5, `form-control` was converted into
`form-select` to `select` tag. This made was due to a caret missmatch
with the previous version due to a change in Boostrap 5.

But in some case the `select` input used Select2. Actually our Select2
is an old version (3.5.4 the last release is 4.0.13) and we have a `CSS`
file alongside Select2 to match the style of Boostrap, but it is for
Boostrap 3. We restore the old class in these case to restore the old
styling (with little hiccup but better than actual).

The good fix, will be to upgrade the lib, and its Bootstrap theming
(for BS5) to the last version.

closes odoo/enterprise#31415

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>