[FIX] account_reports: ellipsis management: don't consider columns with a popup icon as empty
When rendering the report, we remove some empty column values in order to let space to stretch the line name further. When doing that, we didn't check whether the line contained a popup icon or not (for carryover or cell edition, for example), causing them to disappear on 0-balanced leading cells.

closes odoo/enterprise#31429

Signed-off-by: Laurent Smet <las@odoo.com>