[IMP] stock_barcode*: grouped line autoscroll
The autoscroll when a line is selected was pretty annoying with grouped
lines especially when there is a lot of sublines.
To improve that, this commit makes these two changes:

- When a grouped line is selected, it's no more automatically unfolded
  (but conversely, unfolding a grouped line still select it);

- The parent line of a group (the summary) is sticky on the top of the
  page while scrolling through the sublines (that way, the user can
  always see the summary if there it too many lines);

Some changes in the CSS style and on the way the scroll position is
computed were done to match those changes.

task-2960938

X-original-commit: 124edb8f57f4476f4f4aac9fc6a38922d87faaab
Part-of: odoo/enterprise#32481