[IMP] knowledge: Migrate /index /outline
This commit migrates the ``/index`` and ``/outline`` commands to the new Command
system. These two commands work as they used to.

  ``/index`` allows users to list direct children of the current article;

  ``/outline`` allows users to list all descendants of the current article;

When clicking on a link, users are redirected to the related article;

No functional change should occur with this commit as it is mainly a code
migration.

See https://github.com/odoo/enterprise/pull/30270

Task-2796156

Part-of: odoo/enterprise#29423
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>