[IMP] stock_barcode: highlight line's src location
When a source location is scanned, if a line has this location, the
application will scroll to this line and the location will be written in
bold to highlight it.
But it seems it's not enough, so to highlight even more the line source
location, a colored (green) background is added.

task-2960938

closes odoo/enterprise#30979

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>