[IMP] hr_payroll_*: compute dates according to pay schedule
Before this commit, creating a payslip would always compute
the duration of it on one month, even though the pay schedule
defined in the payroll structure says otherwise.
After the commit, setting the employee for the payslip will adapt
the dates accordingly so that the duration is correct.
This commit also adds a warning message if the payslip duration
isn't the right one.
This change affects the fields dependency as `date_from` and
`date_to` now depend on the structure selected for the payslip.

*:
- hr_contract_salary_payroll
- hr_payroll
- l10n_be_hr_payroll
- test_l10n_be_hr_payroll_account

task-2977799

closes odoo/enterprise#33250

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>